Finishing up the merge
